Refactor by-law amendement voting thresholds
Per @roughani's law-knowledgable friend, it was suggested that we make by-law
amendments require a supermajority. After some discussion, we decided that if
we did that, we would be making voting-rights affecting bylaw changes _easier_
to enact that normal bylaw amendments.

To correct this weight change, we decided to allow bylaw amendment votes to
require a 2/3 vote of the board member is in attendance at the meeting
(quorum).

Voting-rights amendments require a 2/3 vote of _all_ board members, regardless
of wether they are attending the meeting or not.

This makes it more difficult to pass a voting-right amendment which is the
balance we were looking for.